When using recvmessage only parse the actual number of pkgs received
The previous code would, when multimsg was configured as 8, always loop
over all 8 package buffers, even if we just received a single new
package.

This will break when on a previous iteration, the buffer was modified in
statsd_packet_parse by injecting a null-byte to terminate the key
string, and now, when erroneously parsing the packet again, choking on
the unexpected null-byte.